Refactor: Cars 일급 객체 적용 수정
